Translate block control flow and breaks
Add support for `block`, `loop`, `br`, `br_if`, `select`. The `br_if`
and `select` instructions benefit from optional fusing with a preceding
condition instruction. The general strategy taken for breaking out of a
block with return values is to stash those values in temporary locals,
unwind the stack back to the right height, then unstash the values (and
kill the locals).