Multi-crate project, introduce `wast2jar`
Crate changes
=============

Switches from one cargo crate to 3:

  - the library (small dependency footprint, contains core translation)
  - the `wasm2jar` binary (command line parser, extra niceties)
  - the `wast2jar` testing binary

The main motivations for the multi-crate approach were: wanting to make
sure it would be possible to have minimal dependencies for the core library
and creating a more flexible testing driver (I mucked around with a
`build.rs` prototype that auto-generated `#[test]` functions, but that
had bad UX and was running into all manner of `cargo` issues.

WAST tester
===========

This is a binary program that can easily run tests on directories full
of WAST files. The eventual target is to make testing against the spec's
WAST files a matter of: cloning the spec repo and running `wast2jar` on
the `spec/test/core` folder.

Internally, each test parses the WAST into:

  - Modules that get translated and either saved as `.class` files or
    assert expected translation errors.
  - Assertions that get put into `JavaHarnessX.java` files. These files
    get run through `javac` and then `java`. They reference the
    generated classes.

All the intermediate outputs of a WAST get put into specific directory
(so they can easily be tweaked/debugged).