Support `br_table`, preliminary renamer
Add support for translating `br_table` into `lookuptable`. The main
optimization/challenge here was avoiding intermediate jump blocks from
the `lookuptable` when they are not necessary. The only time we really
need an intermediate jump block is if there is stack unwinding to be
done before jumping to the final target (since the stack unwinding code
has to be run before the jump).

Also added a very preliminary renaming strategy option. The motivation
for now is just to allow `wast2jar` to handle test cases that have
dashes in the name.