Support multiple returns
Miscallaneous bugs and features:

  * fixed incorrect widths for a bunch of instructions
  * add support for branching out of the entire function
  * deduplicate `br_table` indirect blocks
  * support `assert_return` with multiple returns
  * support `assert_exhaustion`

Multiple returns
================

Functions with multiple return values get translated into functions that
return an array of objects. This means that primitive values get packed
into their boxed variants (and unpacked again when calling such functions).

In order to test multiple returns, a very hacky implementation of `call`
has also been added. There are TODOs in the code documenting the
shortcuts taken.

Control flow and dead code
=========================

Improved logic to avoid generating dead bytecode. The biggest part of
this is that we skip translating operators if there is no current frame.
This doesn't matter for instructions like `iadd` (which anyways get
skipped if there is no current block), but it does matter if the codegen
places a label.

Now, we detect when instructions (as well as entire nested blocks) are
dead. The second important change is that when we place an end label, we
match on the error about never having seen this label and use this as a
signal that the label is unreachable.