Add types representing JVM names
Internally, these are still the same `Cow<'static, str>` (although I did
unsuccessfully try to switch that too `&'a str`). However, the type is
opaque and can only be constructed through code that will validate that
the name is well-formed. Most importantly, it is now much clearer when
we are dealing with short class names vs. fully qualified ones.