Support internal WASM memories
Add support for WSM memory instructions that touch memories that are
neither imported not exported. The implementation approach is to put
these memories inside local fields of type `ByteBuffer`, since the API
for byte buffer is pretty close to what WASM expects.

Active data sections are used in the constructor to initialize the
memory (the actual constant bytes get stored in string constants in the
classfile).