Support `call_indirect`
Support for `call_indirect` works through `invokedynamic`. The appraoch
is documented in more detail beside the codegen of the bootstrap method.

Since I plan to use this approach for more instructions that are grouped
around specific tables or memories, a `BootstrapUtilities` type has been
added to manage a cache of which bootstrap methods have been instantiated.