Implement `memory.grow` and `memory.size`
This is done using a new bootstrap method that gets instantiated once
per memory (same approach as with tables).

Also improved the resolution of methods and fields (ie. for the
non-explicit builder methods) to `class_graph`. Now, rather than just
look on the current class, the resolution will attempt to travel up the
inheritance graph too.