Support importing functions
For now, these are still the only types of entities that can be
imported. The signature of the constructor has been updated to accept a
map of functions to import.

Imported functions get stored as method handle fields with get
`invokeExact`-ed in a trampoline static method.

The `wast2jar` utility is updated to support `register` commands too.