Import/exports using runtime types
This adds a handful of runtime types representing entities that can be
imported/exported. For now, these types are always re-generated, though
in the future they should be pulled out into a re-usable external JAR.

Imports/exports for functions, tables, globals, memories all work
uniformly: if the entity is imported/exported, it uses the boxed
representation.