Support `table`/`memory` `copy`/`init` instructions
Add proper support for element and data sections. Both of these are
lazily intialized and stored on fields in the main WASM module. Prior to
intialization, those fields are `null`. After being dropped, in
accordance with the spec, the fields become empty arrays/buffers. The
lazy intialization is performed by a getter method.

Add support for more bulk operations:

  - `table.init` and `table.copy` call out to `System.arraycopy`
  - `memory.init` and `memory.copy` use `ByteBuffer.put`
  - `data.drop` and `elem.drop` just set data/element fields to `null`

Misc fixes to the harness:

  - accepts confusing unicode (though will now generate bogus Java
    source code)
  - support `assert_unlinkable`
  - more `spectest` imports are specified