Bootstrap methods for external tables/memories
The bootstrap methods for various table and memory operations now work
on external tables and memories as well as internal ones. This was
fairly straightforward - the new external variants of the bootstrap
methods just need to compose getters/setters with a getter for the
actual table or memory object.

A couple other bugs were uncovered:

  - function tables still contained methods whose types had the current
    module as their last argument (eg. how static methods on the module
    work). Fixing this makes it possible to `call_indirect` imported
    functions.

  - Bootstrap methods were not being reused! This looks like a simple
    oversight, easily fixed.

  - Global variable mutability was not checked on module initialization.